Move move calculation functions to new module
Moved move calculation functions to a new module.
